pyproject.toml
[tool.poetry]
name = "tg-msg-sender"
version = "0.1.0"
description = ""
authors = ["Ivan Khimich <himivan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asyncio = "^3.4.3"
python-telegram-bot = "^20.6"

[tool.poetry.scripts]
send-tg-msg = "tg_msg_sender.sender:main"
tg-utility = "tg_msg_sender.utilities:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
